@@ -10,7 +10,7 @@ edit_on_github = True
 [options]
 zip_safe = False
 packages = find:
-python_requires = >=3.7
+python_requires = >=3.8
 setup_requires = setuptools_scm
 install_requires =
     numpy
@@ -23,7 +23,7 @@ install_requires =
     cobaya
     pyccl
     sacc
-    fgspectra @ git+https://github.com/simonsobs/fgspectra@act_sz_x_cib#egg=fgspectra
+    fgspectra
     mflike @ git+https://github.com/simonsobs/lat_mflike@master
 
 [options.package_data]
